# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EasyStay Documentation",
        "description": "Ceci est la documentation de EasyStay.",
        "contact": {
            "email": "haroldavademe0@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/chambre/type/store": {
            "put": {
                "tags": [
                    "Type chambre"
                ],
                "summary": "Créer un type de chambre",
                "operationId": "3e967fe51dd670f0c2f7f14f4994fcb9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "L'ID du type de chambre à mettre à jour.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Chambre Standard"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "features": {
                                        "type": "string",
                                        "example": "Vue sur la mer, Lit King Size"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Le type de chambre a été créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Type de chambre créé avec succès"
                                        },
                                        "type_chambre": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Chambre Standard"
                                                },
                                                "capacity": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "features": {
                                                    "type": "string",
                                                    "example": "Vue sur la mer, Lit King Size"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "AVAILABLE"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé pour les utilisateurs non administrateurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Accès Refusé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "1062": {
                        "description": "Type de chambre déjà existant",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Type de chambre existant."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Type chambre"
                ],
                "summary": "Créer un type de chambre",
                "operationId": "e9c3dfba7b945768ef769e58db595f1f",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nom du type de chambre",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Chambre Standard"
                        }
                    },
                    {
                        "name": "capacity",
                        "in": "query",
                        "description": "Capacité d'accueil de la chambre",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    },
                    {
                        "name": "features",
                        "in": "query",
                        "description": "Caractéristiques supplémentaires de la chambre",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Vue sur la mer, Lit King Size"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Chambre Standard"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "features": {
                                        "type": "string",
                                        "example": "Vue sur la mer, Lit King Size"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Le type de chambre a été créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Type de chambre créé avec succès"
                                        },
                                        "type_chambre": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Chambre Standard"
                                                },
                                                "capacity": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "features": {
                                                    "type": "string",
                                                    "example": "Vue sur la mer, Lit King Size"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "AVAILABLE"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé pour les utilisateurs non administrateurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Accès Refusé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "1062": {
                        "description": "Type de chambre déjà existant",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Type de chambre existant."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/chambre/type/show": {
            "get": {
                "tags": [
                    "Type chambre"
                ],
                "summary": "Afficher les types de chambre",
                "operationId": "c06451e6ee4fd3483075e74838ce6cd1",
                "responses": {
                    "200": {
                        "description": "Liste des types de chambre de l'hôtel",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des types de chambre de votre hôtel"
                                        },
                                        "typeChambre": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Chambre Standard"
                                                    },
                                                    "capacity": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "features": {
                                                        "type": "string",
                                                        "example": "Vue sur la mer, Lit King Size"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "AVAILABLE"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé pour les utilisateurs non administrateurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Accès Refusé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}